Move log enums to avoid tests importing from /extension
