+++ src/extension/analysis/analyzer_status_reporter.ts
@@ -2,11 +2,11 @@ import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
 import { env, ProgressLocation, version as codeVersion, window, workspace } from "vscode";
+import { LogCategory } from "../../shared/enums";
 import { PromiseCompleter } from "../../shared/utils";
 import { WorkspaceContext } from "../../shared/workspace";
 import { Analytics } from "../analytics";
 import { config } from "../config";
-import { LogCategory } from "../debug/utils";
 import { extensionVersion, getRandomInt, getSdkVersion, isStableSdk } from "../utils";
 import { logError } from "../utils/log";
 import { RequestError, ServerErrorNotification, ServerStatusNotification } from "./analysis_server_types";
