+++ src/extension/commands/sdk.ts
@@ -4,13 +4,13 @@ import * as os from "os";
 import * as path from "path";
 import * as vs from "vscode";
 import { ProgressLocation, Uri, window } from "vscode";
+import { LogCategory, LogSeverity } from "../../shared/enums";
 import { Sdks } from "../../shared/interfaces";
 import { flatMap, PromiseCompleter } from "../../shared/utils";
 import { fsPath } from "../../shared/vscode/utils";
 import { WorkspaceContext } from "../../shared/workspace";
 import { config } from "../config";
 import { stripMarkdown } from "../dartdocs";
-import { LogCategory, LogSeverity } from "../debug/utils";
 import { FlutterCapabilities } from "../flutter/capabilities";
 import { FlutterDeviceManager } from "../flutter/device_manager";
 import { getWorkspaceProjectFolders, locateBestProjectRoot } from "../project";
