+++ src/extension/debug/flutter_web_run.ts
@@ -1,5 +1,5 @@
+import { LogSeverity } from "../../shared/enums";
 import { FlutterRunBase, RunMode } from "./flutter_run_base";
-import { LogSeverity } from "./utils";
 
 export class FlutterWebRun extends FlutterRunBase {
 	constructor(mode: RunMode, pubBinPath: string, projectFolder: string, args: string[], envOverrides: any, logFile: string, logger: (message: string, severity: LogSeverity) => void, maxLogLineLength: number) {
