+++ src/extension/extension.ts
@@ -3,9 +3,12 @@ import * as path from "path";
 import { isArray } from "util";
 import * as vs from "vscode";
 import { flutterExtensionIdentifier } from "../shared/constants";
+import { LogCategory } from "../shared/enums";
 import { Sdks } from "../shared/interfaces";
 import { internalApiSymbol } from "../shared/symbols";
+import { dartPlatformName, forceWindowsDriveLetterToUppercase, isWin, isWithinPath, platformDisplayName } from "../shared/utils";
 import { InternalExtensionApi } from "../shared/vscode/interfaces";
+import { fsPath } from "../shared/vscode/utils";
 import { Context } from "../shared/vscode/workspace";
 import { WorkspaceContext } from "../shared/workspace";
 import { Analyzer } from "./analysis/analyzer";
@@ -70,9 +73,6 @@ import * as util from "./utils";
 import { addToLogHeader, clearLogHeader, getExtensionLogPath, log, logError, logTo } from "./utils/log";
 import { DartPackagesProvider } from "./views/packages_view";
 import { TestItemTreeItem, TestResultsProvider } from "./views/test_view";
-import { LogCategory } from "./debug/utils";
-import { fsPath } from "../shared/vscode/utils";
-import { forceWindowsDriveLetterToUppercase, dartPlatformName, platformDisplayName, isWin, isWithinPath } from "../shared/utils";
 
 const DART_MODE: vs.DocumentFilter = { language: "dart", scheme: "file" };
 const HTML_MODE: vs.DocumentFilter = { language: "html", scheme: "file" };
