+++ src/extension/sdk/dev_tools.ts
@@ -4,13 +4,12 @@ import * as path from "path";
 import * as vs from "vscode";
 import { window, workspace } from "vscode";
 import { CHROME_OS_DEVTOOLS_PORT, pleaseReportBug } from "../../shared/constants";
-import { FlutterService } from "../../shared/enums";
+import { FlutterService, LogCategory } from "../../shared/enums";
 import { Sdks } from "../../shared/interfaces";
 import { isChromeOS } from "../../shared/utils";
 import { Analytics } from "../analytics";
 import { DebugCommands, debugSessions } from "../commands/debug";
 import { config } from "../config";
-import { LogCategory } from "../debug/utils";
 import { PubGlobal } from "../pub/global";
 import { StdIOService, UnknownNotification } from "../services/stdio_service";
 import { getRandomInt } from "../utils";
