+++ src/test/debug_helpers.ts
@@ -3,9 +3,9 @@ import { ChildProcess } from "child_process";
 import { DebugConfiguration, Uri } from "vscode";
 import { DebugProtocol } from "vscode-debugprotocol";
 import { ObservatoryConnection } from "../extension/debug/dart_debug_protocol";
-import { LogCategory, LogSeverity } from "../extension/debug/utils";
 import { log, logProcess } from "../extension/utils/log";
 import { safeSpawn } from "../extension/utils/processes";
+import { LogCategory, LogSeverity } from "../shared/enums";
 import { isWin } from "../shared/utils";
 import { DartDebugClient } from "./dart_debug_client";
 import { defer, getLaunchConfiguration } from "./helpers";
