+++ src/test/flutter_create_tests/flutter_create.test.ts
@@ -2,9 +2,9 @@ import * as assert from "assert";
 import * as fs from "fs";
 import * as path from "path";
 import * as vs from "vscode";
-import { LogCategory, LogSeverity } from "../../extension/debug/utils";
 import { log } from "../../extension/utils/log";
 import { dartCodeExtensionIdentifier } from "../../shared/constants";
+import { LogCategory, LogSeverity } from "../../shared/enums";
 import { internalApiSymbol } from "../../shared/symbols";
 import { InternalExtensionApi } from "../../shared/vscode/interfaces";
 import { fsPath } from "../../shared/vscode/utils";
