+++ src/test/helpers.ts
@@ -5,7 +5,6 @@ import * as path from "path";
 import * as semver from "semver";
 import * as sinon from "sinon";
 import * as vs from "vscode";
-import { LogCategory, LogSeverity } from "../extension/debug/utils";
 import { DelayedCompletionItem } from "../extension/providers/dart_completion_item_provider";
 import { isAnalyzable, vsCodeVersionConstraint } from "../extension/utils";
 import { tryDeleteFile } from "../extension/utils/fs";
@@ -14,7 +13,7 @@ import { waitFor } from "../extension/utils/promises";
 import { PackageDep } from "../extension/views/packages_view";
 import { SuiteTreeItem } from "../extension/views/test_view";
 import { dartCodeExtensionIdentifier } from "../shared/constants";
-import { TestStatus } from "../shared/enums";
+import { LogCategory, LogSeverity, TestStatus } from "../shared/enums";
 import { internalApiSymbol } from "../shared/symbols";
 import { flatMap } from "../shared/utils";
 import { InternalExtensionApi, TestItemTreeItem, TestResultsProvider } from "../shared/vscode/interfaces";
