+++ src/test/mocha_logging_reporter.ts
@@ -1,6 +1,6 @@
 import { ITest, reporters } from "mocha";
-import { LogCategory, LogSeverity } from "../extension/debug/utils";
 import { log } from "../extension/utils/log";
+import { LogCategory, LogSeverity } from "../shared/enums";
 import testRunner = require("vscode/lib/testrunner");
 
 export class LoggingReporter extends reporters.Base {
